refactor(plugin)!: unify managed lifecycle and pipeline configuration
Unify plugin lifecycle, configuration precedence, pipeline chaining, cancellation, and failure handling. Include regression coverage, CLI streaming shutdown fixes, and source-verified architecture diagrams.

Make the log output plugin own the single logging threshold for event output and application diagnostics. Default to log at info, support errors-only logging, and remove the separate application logging setting and CLI none output.

Verify transaction, governance, and rollback log delivery. Add focused local CLI integration tests for output routing, configuration, filters, startup failures, RPC recovery, and shutdown. Update configuration examples, migration guidance, and tray output selection.

+++ event/governance_test.go
@@ -182,7 +182,7 @@ func TestExtractUpdateCommitteeAction(t *testing.T) {
 			Credentials: []common.Credential{
 				cred1,
 			},
-			CredEpochs: map[*common.Credential]uint{
+			CredEpochs: map[*common.Credential]uint64{
 				cred2: 500,
 				cred3: 600,
 			},
